make it work

+++ drainify/record.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python2
 # encoding: utf-8
 from __future__ import print_function
 
@@ -74,6 +74,7 @@ class Recorder(object):
         length = metadata['Metadata']["mpris:length"]
         # avoid recording the beginning of the next track
         milli_secs = length * 1E-3 - 750
+        self.finished_ts = time.time() + length / 1E6
 
         # handler that stops recording of track
         self.timeout_handler = gobject.timeout_add(int(milli_secs),
@@ -81,32 +82,50 @@ class Recorder(object):
 
     def stop_handler(self):
         """ Killing running recorders immediately """
-        running_recs.pop(self.recorder.pid)
-        os.killpg(self.recorder.pid, signal.SIGKILL)
-
-        # remove temporary file
-        os.remove(self.tmp_name)
+        print("%d/%d  %s  stop_handler"%(self.recorder.pid,self.encoder.pid,self.metadata['Metadata']['xesam:title']))
+        print("should stop at %f, now its %f"%(self.finished_ts, time.time()))
 
         # remove timeout handler
         gobject.source_remove(self.timeout_handler)
+
+        if time.time() + 5 >= self.finished_ts:
+            print("close enough, calling regular stop")
+            self.stop_recording_cb()
+        else:
+            print("seems to be a skip, killing old recording")
+
+            running_recs.pop(self.recorder.pid)
+            os.killpg(self.recorder.pid, signal.SIGKILL)
+
+            # remove temporary file
+            os.remove(self.tmp_name)
+
+    def debugout(self, s):
+        print("%d/%d  %s  %s"%(self.recorder.pid,self.encoder.pid,self.metadata['Metadata']['xesam:title'], s))
 
     def stop_recording_cb(self):
         """ Callback for stopping the recording.
         Setting ID3 Tags.
         Moves the temp file to the specific directory.
         """
+        self.debugout("stop_recording_cb (timeout) ")
         running_recs.pop(self.recorder.pid)
         os.killpg(self.recorder.pid, signal.SIGKILL)
 
+        self.debugout("recorder killed")
         title = self.metadata['Metadata']['xesam:title']
         artist = self.metadata['Metadata']['xesam:artist'][0]
 
         final_name = rec_dir + '/%s - %s.mp3' % (artist, title)
+        self.debugout("final_name = "+final_name)
 
         # waiting for encoder to encode the end of stream
+        self.debugout("waiting for encoder to exit...")
         self.encoder.wait()
+        self.debugout("ok")
 
         set_id3_tags(self.tmp_name, self.metadata['Metadata'])
+        self.debugout("id3 ok")
 
         # moving temp recording to destination then finished
         shutil.move(self.tmp_name, final_name)
@@ -116,10 +135,11 @@ class Recorder(object):
 
 # TODO: skipping tracks dont work
 def recording_handler(sender=None, metadata=None, sig=None):
+    print("handler",metadata)
     if "PlaybackStatus" in metadata:
         if metadata['PlaybackStatus'] == 'Paused':
             # dont stop befor last song recording ended
-            time.sleep(10)
+            time.sleep(1.5)
             for pid, rec in running_recs.items():
                 rec.stop_handler()
             return
@@ -132,6 +152,8 @@ def recording_handler(sender=None, metadata=None, sig=None):
             else:
                 # message contains metadata and PlaybackStatus means track skipping
                 for pid, rec in running_recs.items():
+                    if rec.metadata['Metadata']['mpris:trackid'] == metadata['Metadata']['mpris:trackid']:
+                        return
                     rec.stop_handler()
 
     title = metadata['Metadata']['xesam:title']
@@ -140,16 +162,17 @@ def recording_handler(sender=None, metadata=None, sig=None):
     print("recording: %s - %s" % (artist, title))
 
     # not record the end of the last song, so sleep
-    time.sleep(1.0)
+    #time.sleep(1.0)
 
     parec = subprocess.Popen(['parec',
-                              '--format=', 's16le',
-                              '--rate=', '44100',
-                              '--device=', 'combinded.monitor'
+                              '--format=s16le',
+                              '--rate=44100',
+                              '--device=SpotifyGrabberSink.monitor',
+                              '--record',
                               ],
                               stdout=subprocess.PIPE,
                               preexec_fn=os.setsid,
-                              shell=True
+                              #shell=True
                           )
 
     lame = subprocess.Popen(['lame',
@@ -206,24 +229,30 @@ def main():
         global rec_dir
         rec_dir = os.path.abspath(args.dir)
 
+
+    # create loopback sink if not exists
+    subprocess.check_call("pactl list sinks | grep SpotifyGrabberSink || pactl load-module module-null-sink sink_name=SpotifyGrabberSink", shell=True)
+
     # init combined sink
     sinks = pa.list_sinks()
 
     sink_choose = None
 
-    if len(sinks) > 1:
-        for i, s in enumerate(sinks):
-            print("%i: %s" % (i, s))
+    #if len(sinks) > 1:
+    #    for i, s in enumerate(sinks):
+    #        print("%i: %s" % (i, s))
+    for i, s in enumerate(sinks):
+        if s == "SpotifyGrabberSink": sink_choose = i
 
-        sink_choose = raw_input("Choose your audio device (Default [0]): ")
+     #   sink_choose = raw_input("Choose your audio device (Default [0]): ")
 
     # default sink
-    if not sink_choose:
-        sink_choose = 0
+    #if not sink_choose:
+    #    sink_choose = 0
 
-    rec_sink = sinks[sink_choose]
+    rec_sink = sinks[int(sink_choose)]
     spot_id = pa.find_spotify_input_sink()
-    combined_sink = pa.create_combined_sink(rec_sink)
+    #combined_sink = pa.create_combined_sink(rec_sink)
     pa.move_sink_input(spot_id)
 
     dbus.mainloop.glib.DBusGMainLoop(set_as_default=True)
@@ -245,7 +274,7 @@ def main():
         loop.run()
     except KeyboardInterrupt:
         print("Received KeyboardInterrupt. Quiting.")
-        pa.unload_combined_sink(combined_sink)
+        #pa.unload_combined_sink(combined_sink)
         cleanup()
         sys.exit()
